# vafund-backend/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend API untuk aplikasi VaFund menggunakan Go Fiber dan Hyperledger Fabric",
        "title": "VaFund Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@vafund.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/api/events": {
            "get": {
                "description": "Get a list of all events",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Get all events",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.EventResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new event with code, name, start date, end date and active status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Create a new event",
                "parameters": [
                    {
                        "description": "Event data",
                        "name": "event",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateEventRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.EventResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/events/active": {
            "get": {
                "description": "Get a list of events that are currently active and within their date range",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Get active events",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.EventResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/events/{code}": {
            "get": {
                "description": "Get a specific event by its code",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Get event by code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event Code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.EventResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update event details including name, dates, and status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Update event details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event Code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Event update data",
                        "name": "event",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateEventDetailRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.EventResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/events/{code}/status": {
            "put": {
                "description": "Update the active status of an event (activate/deactivate)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Update event status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event Code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Event status update",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateEventRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.EventResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CreateEventRequest": {
            "type": "object",
            "required": [
                "code",
                "description",
                "endDate",
                "isActive",
                "name",
                "startDate"
            ],
            "properties": {
                "code": {
                    "type": "string",
                    "example": "RAMADAN2025"
                },
                "description": {
                    "type": "string",
                    "example": "Annual charity drive during Ramadan month"
                },
                "endDate": {
                    "type": "string",
                    "example": "2025-04-30T23:59:59Z"
                },
                "isActive": {
                    "type": "string",
                    "example": "true"
                },
                "name": {
                    "type": "string",
                    "example": "Ramadan Charity Drive"
                },
                "startDate": {
                    "type": "string",
                    "example": "2025-03-01T00:00:00Z"
                }
            }
        },
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Detailed error description"
                },
                "message": {
                    "type": "string",
                    "example": "Error message"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "models.EventResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Event created successfully"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "models.UpdateEventDetailRequest": {
            "type": "object",
            "required": [
                "description",
                "endDate",
                "isActive",
                "name",
                "startDate"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Updated description for annual charity drive"
                },
                "endDate": {
                    "type": "string",
                    "example": "2025-04-30T23:59:59Z"
                },
                "isActive": {
                    "type": "string",
                    "example": "true"
                },
                "name": {
                    "type": "string",
                    "example": "Updated Ramadan Charity Drive"
                },
                "startDate": {
                    "type": "string",
                    "example": "2025-03-01T00:00:00Z"
                }
            }
        },
        "models.UpdateEventRequest": {
            "type": "object",
            "required": [
                "isActive"
            ],
            "properties": {
                "isActive": {
                    "type": "string",
                    "example": "false"
                }
            }
        },
...
    }
}